EXPERIMENT: In this checkin I refactored and cleaned up the logging. An EVENTLOG env var now can redirect the logging output to either the binary .eventlog file or use textual stderr print messages.
The results are odd.

  * Failed twice on 100X runs VERBOSITY=3, textual logging
    (at least one was the resource exhausted problem)
  * Succeeded once in the same config
  * Succeeded on two 100X runs with binary logging

However, when I up the size to fib(20) I get the resource exhausted
error with successful VERBOSITY=3 binary event logging.  It looks like
there was a stealrequest and no steal response.  However, the thread
involved (thread 8) seems to have blocked on an mvar and then gone on
to be rescheduled again and again. There's nothing about exceptions
in the eventlog.