Added a backoff resource and attached it to Dist.hs
This does not eliminate the resource vanished bug.  And
it's hard to say how it changes frequency (with the pipes transport).

A run of 100X fib(15)'s passed, but then, under VERBOSITY=0
fib(16) crashed on iteration 21 and VERBOSITY=4 crashes quicker!
(iteration 1, iteration 3)

And as has been the case, a big enough run (e.g. fib(21)) will go wrong
on the first try.