Added code to make IVar UIDs *actually* guaranteed to be unique.
Passes fib(22) via run_two_local with both TCP and Pipes.  Hmm.
Both of them are very slow (~35 seconds), in both cases the processes
are quite lazy, using 3% and 20% CPU.  In the pipes version the warnings
indicate that reading the pipe (presumably for the next message) took
as much as 14 seconds.

Note that it performs similarly even when -eventlog / -l is turned OFF and VERBOSITY=0.