Small addition: an extra guard at shutdown time to make sure there are ZERO outstanding longspawn-related IVars.
Hmm. I've made a number of refactorings and misc fixes recently and
now I'm having trouble getting my beloved bug!  fib(22) still passes
on both pipes/tcp.

   * rep 100 ./run_two_local.sh tcp 10 +RTS -N1
   * rep 100 ./run_two_local.sh pipes 10 +RTS -N1
   * rep 100 ./run_two_local.sh tcp 10 +RTS -N4

I did see a failure with this command:

   rep 100 ./run_two_local.sh pipes 10 +RTS -N4

Specifically, it produced the following messages in a loop:

    WARNING: have been trying an IO action (read pipe file descriptor) for 3.036801 seconds without success
    WARNING: have been trying an IO action (open pipe for write) for 3.033006 seconds without success
    WARNING: have been trying an IO action (read pipe file descriptor) for 4.057738 seconds without success
    WARNING: have been trying an IO action (open pipe for write) for 4.051932 seconds without success

And there's NO pipe_* files to be found in /tmp/ in that case!  It
looks like something went wrong in the multithreaded case with setting
up the pipes (data race?)

The following one, however, failed on iteration 13 with an exception!!

    rep 100 ./run_two_local.sh tcp 15 +RTS -N4