Fix outdated links in the doc of  `Control.Monad.Par` (#78)
- For the paper,

use the `Links` markup instead of just `URL` markup,
add the paper's info with IEEE citation style,
replace the outdated link with the Mr.Marlow's pasted blog on github.io.

- For the slides,
replace the outdated link with the link of a copy of source file in cse.unt.edu ,
remove the invalid link which seems cannot find the existed source.